feat: add local dummy db
- delete redundant profile validators
