+++ src/server/auth/provider/index.ts
@@ -1,3 +1,3 @@
-export { default as github } from './github/strategy';
-export { default as google } from './google/strategy';
+export { default as github } from './github';
+export { default as google } from './google';
 export { default as local } from './local';
