+++ src/server/auth/provider/local.ts
@@ -1,5 +1,30 @@
-import Local from 'passport-local';
+import { Profile } from 'passport';
+import { Strategy as LocaLStrategy } from 'passport-local';
 
-export default new Local.Strategy((username, password, done) => {
-  done(null, { username, password });
+interface LocalDBDataset extends Omit<Profile, 'provider'> {
+  password: string;
+}
+
+const localDB: LocalDBDataset[] = [
+  {
+    username: 'ljosberinn',
+    password: 'next-with-batteries!',
+    id: '1',
+    displayName: 'gerrit alex',
+  },
+];
+
+export default new LocaLStrategy((username, pw, done) => {
+  // custom logic for your db goes here!
+  const user = localDB.find(
+    user => user.username === username && user.password === pw
+  );
+
+  if (!user) {
+    return done(null);
+  }
+
+  const { password, ...response } = user;
+
+  done(null, { ...response, provider: 'local' });
 });
